# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CERTIFICATION-SIMPLON",
        "version": "1.0"
    },
    "paths": {
        "/api/annonces": {
            "get": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Liste de toutes les annonces",
                "description": "Récupère la liste de toutes les annonces non encore prises en charge.",
                "operationId": "getAnnonces",
                "responses": {
                    "200": {
                        "description": "Liste des annonces récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste de toutes les annonces"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur s'est produite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ajoutAnnonce": {
            "post": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Créer une nouvelle annonce",
                "description": "Permet à un étudiant de créer une nouvelle annonce.",
                "operationId": "createAnnonce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Description de l'annonce"
                                    },
                                    "budget": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    },
                                    "caracteristiques": {
                                        "type": "string",
                                        "example": "Caractéristiques de l'annonce"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Annonce créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Annonce créée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation a échoué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données de la requête sont invalides."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "description": [
                                                    "Le champ description est requis."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/detailAnnonce/{id}": {
            "get": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Obtenir une annonce par son identifiant",
                "description": "Récupère une annonce spécifique en utilisant son identifiant.",
                "operationId": "getAnnonceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'annonce",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Annonce récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Annonce non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Annonce non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Une erreur s'est produite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur s'est produite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/annonces/{annonce}": {
            "put": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Mettre à jour une annonce",
                "description": "Met à jour une annonce spécifique en utilisant son identifiant.",
                "operationId": "updateAnnonce",
                "parameters": [
                    {
                        "name": "annonce",
                        "in": "path",
                        "description": "Objet de l'annonce à mettre à jour",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "budget": {
                                        "type": "number"
                                    },
                                    "caracteristiques": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Annonce mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Annonce mise à jour avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Vous n'avez pas la permission de mettre à jour cette annonce",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas la permission de mettre à jour cette annonce"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Une erreur s'est produite lors de la mise à jour de l'annonce",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur s'est produite lors de la mise à jour de l'annonce"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/annonces/{id}": {
            "delete": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Supprimer une annonce",
                "description": "Supprime une annonce spécifique en utilisant son identifiant.",
                "operationId": "deleteAnnonce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'annonce à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Annonce supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Annonce supprimée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Vous n'avez pas la permission de supprimer cette annonce",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas la permission de supprimer cette annonce"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Annonce non supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Annonce non supprimée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}